backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pinecone
numpy
PyMuPDF
python-multipart
packaging
